refactor(base): Rename ExternalRequest to Requests
Reasons:
    - Make this utility class like the other (ending with an "s")
    - This class can be used for all requests (internal or external)
