+++ addons/osio-addon/src/main/java/io/fabric8/launcher/osio/client/OsioAuthClient.java
@@ -9,8 +9,8 @@ import javax.inject.Inject;
 import io.fabric8.launcher.base.identity.TokenIdentity;
 import okhttp3.Request;
 
-import static io.fabric8.launcher.base.http.ExternalRequest.executeAndParseJson;
-import static io.fabric8.launcher.base.http.ExternalRequest.securedRequest;
+import static io.fabric8.launcher.base.http.Requests.executeAndParseJson;
+import static io.fabric8.launcher.base.http.Requests.securedRequest;
 import static io.fabric8.launcher.osio.OsioConfigs.getAuthUrl;
 import static io.fabric8.utils.URLUtils.pathJoin;
 import static java.util.Objects.requireNonNull;
